+++ android/src/main/kotlin/top/kikt/flutter_image_editor/option/FormatOption.kt
@@ -1,0 +1,6 @@
+package top.kikt.flutter_image_editor.option
+
+/// create 2019-10-14 by cai
+
+
+data class FormatOption(val format: Int, val quality: Int)